Add support for hash joins (JoinOperation) to the DSL.  Previously only the slow NestedLoopJoinOperation was supported by the 'join' macro.
A join can now be defined in DSL as:

leftjoin:
   right key1, key2:
      input "CONNECTION1", Command="SELECT..."
   left keyA, keyB:
      input "CONNECTION2", Command="SELECT..."
   action:
      row = left.Clone()
      row.connection1data = right.col1

See InnerHashJoin DSL test for further example.

Additionally added support to pass in additional namespaces when
creating the DSL compiler, and added a method to AbstractClassGenerator<T>
to allow macros to add a method specification, in advance of the actual
class definition being rendered.